Added rest of views, added scan product and add product forms with connection to DataBase.

+++ src/main/java/com/github/sirlacky/DietManager/domain/repositories/ProductRepository.java
@@ -6,7 +6,9 @@ import java.util.Optional;
 
 public interface ProductRepository extends JpaRepository<Product,Long> {
 
-    Optional<Product> findProductByCode(int code);
+    Optional<Product> findProductByCode(Long code);
+
+
 
 }
